Update Supporting_functions_v2.R
Fixed a bug resulting from an update to the output of the kendallTrendTest() function in a recent version of the EnvStats package. The fit object output from the updated function is now classed as an "htestEnvStats object" and is therefore not recognised by broom::tidy(). Forced class change to an "htest object" using structure() to allow broom::tidy() to produce tidy model statistics.